refactor: pkg 디렉토리를 도메인명으로 변경
- strategy/pkg/ → strategy/payment/
- factory/pkg/ → factory/notification/
- package 이름, import 경로, README 파일 구성 테이블 반영
