+++ golang/pattern/factory/notification/domain.go
@@ -1,4 +1,4 @@
-package pkg
+package notification
 
 import "context"
