+++ golang/pattern/factory/notification/email_service.go
@@ -1,4 +1,4 @@
-package pkg
+package notification
 
 import (
 	"context"
