+++ golang/pattern/factory/notification/user_preference_store.go
@@ -1,4 +1,4 @@
-package pkg
+package notification
 
 import (
 	"context"
