+++ golang/pattern/strategy/payment/credit_card_service.go
@@ -1,4 +1,4 @@
-package pkg
+package payment
 
 import (
 	"context"
