+++ golang/pattern/strategy/payment/helper.go
@@ -1,4 +1,4 @@
-package pkg
+package payment
 
 func generateID() string {
 	return "12345678"
