+++ golang/pattern/strategy/payment/kakao_pay_service.go
@@ -1,4 +1,4 @@
-package pkg
+package payment
 
 import (
 	"context"
