+++ golang/pattern/strategy/payment/naver_pay_service.go
@@ -1,4 +1,4 @@
-package pkg
+package payment
 
 import (
 	"context"
