+++ golang/pattern/strategy/payment/payment_usecase.go
@@ -1,4 +1,4 @@
-package pkg
+package payment
 
 import (
 	"context"
